+++ scripts/deploy-md.js
@@ -13,7 +13,8 @@ const version = require("../package.json").version;
 
 const minOptions = {
   removeComments: true,
-  collapseWhitespace: true
+  collapseWhitespace: true,
+  removeAttributeQuotes: true
 };
 
 marked.setOptions({
@@ -22,17 +23,28 @@ marked.setOptions({
   }
 });
 
+let lines;
+
 fs.readFileAsync("index.html", "utf8")
   .then(html => html.replace(/dist\/bundle\.js\?v=\d+\.\d+\.\d+/, `dist/bundle.js?v=${version}`))
   .then(html => html.replace(/dist\/bundle\.css\?v=\d+\.\d+\.\d+/, `dist/bundle.css?v=${version}`))
+  .then(html => htmlmin(html, minOptions))
   .then(html => fs.writeFileAsync("index.html", html, "utf8"))
 ;
 
 glob("./src/md/**/*.md", (err, files) => {
   files.forEach(filepath => {
     fs.readFileAsync(filepath, "utf8")
+    .then(content => {
+      lines = content.split("\n");
+      if(lines.indexOf("---") === 0 && lines.indexOf("---", 1) === 2 && lines[1].indexOf("tags") > -1) {
+        return lines.splice(3).join("\n");
+      } else {
+        return content;
+      }
+    })
     .then(content => marked(content))
-    .then(html => htmlmin(html))
+    .then(html => htmlmin(html, minOptions))
     .then(html => {
       return mkdirp.mkdirpAsync(filepath
         .replace("./src/md/", "./dist/partials/")